refactor: 폴더 구조 변경 및 불필요한 파일 삭제
