--- src/main/java/com/callog/callog_diet/api/open/DietController.java
@@ -1,4 +1,0 @@
-package com.callog.callog_diet.api.open;
-
-public class DietController {
-}
