+++ src/main/java/com/callog/callog_diet/domain/repository/FoodRepository.java
@@ -1,9 +1,7 @@
 package com.callog.callog_diet.domain.repository;
 
-import com.callog.callog_diet.domain.Food;
-
+import com.callog.callog_diet.domain.entity.Food;
 import org.springframework.data.jpa.repository.JpaRepository;
-
 import java.util.List;
 
 public interface FoodRepository extends JpaRepository<Food, Long> {
